ATF-9: Moved EventManger methods into DefaultFeedWriter
